refactor : Controller 부분 작성, App.js에 구동 기능 추가
